refactor ted names
