+++ conf/config.yaml
@@ -3,7 +3,7 @@ database:
 defaults:
   - dataset: sgd.yaml
   - state: ted.yaml
-  - model: ss.yaml
+  - model: ted.yaml
 resources:
   gpus: 0
   seed: 2020
