--- conf/model/ss.yaml
@@ -1,26 +1,0 @@
-n_features: -1
-encoding_dimension: 40
-hidden_layers_sizes_pre_dial: [40]
-hidden_layers_sizes_bot: []
-hidden_layers_sizes_post_dial: []
-heads: 4
-num_layers: 1
-transformer_size: 128
-unidirectional_encoder: true
-embedding_space: 20
-num_neg: 5
-regularization_constant: 1e-3
-dropout_attention: 0.0
-dropout_hidden: 0.0
-dropout_dialogue: 0.1
-dropout_action: 0.0
-batch_size: 64
-similarity: "cosine"
-loss: "margin"
-epochs: 1
-opt: "Adam"
-activation: "relu"
-lr: 1e-3
-state: "binary" # "binary" or "rse"
-class_correction: False
-name: "SS"
