+++ conf/model/ted.yaml
@@ -1,6 +1,6 @@
 n_features: -1
 encoding_dimension: 40
-hidden_layers_sizes_pre_dial: [[-1, 40]]
+hidden_layers_sizes_pre_dial: [40]
 hidden_layers_sizes_bot: []
 hidden_layers_sizes_post_dial: []
 heads: 4
@@ -15,11 +15,12 @@ dropout_hidden: 0.0
 dropout_dialogue: 0.1
 dropout_action: 0.0
 batch_size: 64
-type_batch: "sequence"
+similarity: "cosine"
+loss: "margin"
 epochs: 10
 opt: "Adam"
 activation: "relu"
 lr: 1e-3
-embedding_type: "binary" # "binary" or "rse"
-name: "TED"
+state: "binary" # "binary" or "rse"
 class_correction: False
+name: "TED"
